requirements.txt
openai
pinecone
python-dotenv
pypdf
tiktoken
tqdm 